ダメージを可視化するシステムを追加

+++ Assets/MpScript.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class MpScript : ItemBase
 {
-    [SerializeField] int _mp;
+    [SerializeField] float _mp;
     // Start is called before the first frame update
     private void OnDestroy()
     {
